% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{abrevia_nome_meio}
\alias{abrevia_nome_meio}
\title{Abrevia o nome do meio.}
\usage{
abrevia_nome_meio(base, ..., suffixo = "_abrev")
}
\arguments{
\item{base}{A data table, data frame or character vector.}

\item{...}{columns for apply the function}

\item{suffixo}{A character indicating the final part of the new columns' names}
}
\value{
the base parameter with a new column.
}
\description{
\code{abrevia_nome_meio} return names .
}
\examples{
   base <- data.frame(nome = c("Carlos Pereira Neves", "Pedro Aparecido Anjos"))
   base <- remove_preposicao_nomes(base, "nome")
}
