% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_nomes}
\alias{base_nomes}
\title{Brazilian names}
\format{A data frame with 1100000 rows and 2 variables:
\describe{
  \item{nome}{the name}
  \item{sexo}{the gender of the name}
}}
\usage{
base_nomes
}
\description{
A base with most common Brazilian names and the gender of the name.
}
\keyword{datasets}
