% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caixa}
\alias{caixa}
\title{Generate phonetic versions of strings with Caixa's method}
\usage{
caixa(word, max = 20L)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{max}{maximum length of the resulting encodings, in characters}
}
\value{
a character vector containing the phonetical output of
\code{word}, or an NA if the \code{word} value is NA
}
\description{
The function \code{caixa} phonentically encodes the given input
using algorithm developed by Caixa Economica Federal.
}
\details{
To be inserted
}
\examples{
caixa("")
caixa(c("school", "benji"))

}
