% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contador_letras.R
\name{contar_letras}
\alias{contar_letras}
\title{Remove de da e dos of names .}
\usage{
contar_letras(base, columns, suffix = NULL)
}
\arguments{
\item{base}{A data table, data frame or character vector.}

\item{columns}{for apply the function}

\item{suffix}{Suffix name for the new column.}
}
\value{
a data.table number of any latters on alphabet for each column
}
\description{
\code{contar_letras} return number of any latters on alphabet for each column.
}
\examples{
   base <- data.frame(nome = c("João das Neves", "Pedro dos Anjos", "Maria das Gracas"))
   base <- contar_letras(base, "nome")
}
