% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_nome_proprio.R
\name{extrai_NomeProprio}
\alias{extrai_NomeProprio}
\title{Check Brazilian names}
\usage{
extrai_NomeProprio(x, surname = FALSE, gender = FALSE, stringdist = TRUE,
  spaces = TRUE)
}
\arguments{
\item{x}{List, character or factor with names to be parsed.}

\item{surname}{If TRUE, the list of surnames is returned.}

\item{gender}{If TRUE, the list of gender based on the names is returned.}

\item{stringdist}{if TRUE, make a prediction based on the string distance of Jaro-Winkler between the source data and the input.}

\item{spaces}{if TRUE, returns the names without spaces. If FALSE, it compress all the blank spaces.}
}
\value{
Returns a data.table
}
\description{
\code{extrai_NomeProprio} Parse Brazilian names and returns given names, surnames and gender
}
