% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geocod_base}
\alias{geocod_base}
\title{Brazilian address}
\format{A data frame with 5 rows and 12 variables:
\describe{
  \item{uf}{state of Brazil}
  \item{MatchedAddress}{Address returned from GALILEO}
  \item{cep}{Zip code}
  \item{enderecofinal}{Prompted Address}
  \item{Nome_Municipio}{City names}
}}
\usage{
geocod_base
}
\description{
Some addresses from Brazil
}
\keyword{datasets}
