% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{ident_erros_munic_galileo}
\alias{ident_erros_munic_galileo}
\title{\code{ident_erros_munic_galileo} Returns a new column called munmatch with true or false. This column identify where GALILEO failed}
\usage{
ident_erros_munic_galileo(base, mun, match, uf)
}
\arguments{
\item{base}{Data frame, data set with return from GALILEO}

\item{mun}{character, the name of the municipio.}

\item{match}{character, the colum MatchAdress from GALILEO.}

\item{uf}{character, the name of the state.}
}
\value{
Returns a new column called munmatch with true or false.
}
\description{
\code{ident_erros_munic_galileo} Returns a new column called munmatch with true or false. This column identify where GALILEO failed
}
\examples{
  nome_de_solteira(nome_casada = "Maria Conceicao da Costa", nome_conjuge = "Mario Silva da Costa")
}
