% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lista}
\alias{lista}
\title{Brazilian treatment pronouns}
\format{Character}
\usage{
lista
}
\description{
A regex character with the most common Brazilian treatment pronouns used in Brazil
}
\keyword{datasets}
