% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{remove_preposicao_nomes}
\alias{remove_preposicao_nomes}
\title{Remove de da e dos of names .}
\usage{
remove_preposicao_nomes(base, column, ...)
}
\arguments{
\item{base}{A data table or data frame.}

\item{column}{Caracter, the column with the names.}

\item{...}{more columns for apply the function}
}
\value{
the base param with a new column.
}
\description{
\code{remove_preposicao_nomes} return names without de, da e dos.
}
