% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{remove_pronome_tratamento}
\alias{remove_pronome_tratamento}
\title{Remove commom treatment pronouns used in Brazil.}
\usage{
remove_pronome_tratamento(base, ..., suffixo = "_sem_pron")
}
\arguments{
\item{base}{A data table, data frame or character vector.}

\item{...}{columns for apply the function}

\item{suffixo}{Name of the new column to be created.}
}
\value{
the base param with a new column.
}
\description{
\code{remove_pronome_tratamento} return names without treatment pronouns (Sra, Sr, Dr, etc).
}
\examples{
   remove_pronome_tratamento("Dr. Fulano")
   remove_pronome_tratamento("Exmo. Sr. Cicrano de Tal")

   base <- data.frame(nome = c("Ph.D Pedro dos Anjos", "Prof Maria das Gracas", "Pe. João das Neves"))
   base <- remove_pronome_tratamento(base, "nome", suffixo = "_new_names")

}
