% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{remove_preposicao_nomes}
\alias{remove_preposicao_nomes}
\title{Remove de da e dos of names .}
\usage{
remove_preposicao_nomes(base, ..., suffixo = "_semD")
}
\arguments{
\item{base}{A data table, data frame or character vector.}

\item{...}{columns for apply the function}

\item{suffixo}{Suffix name for the new column.}
}
\value{
the base parameter with a new column.
}
\description{
\code{remove_preposicao_nomes} return names without de, da e dos.
}
\examples{
   base <- data.frame(nome = c("João das Neves", "Pedro dos Anjos", "Maria das Gracas"))
   base <- remove_preposicao_nomes(base, "nome")
}
